replace userType to Integer from Long

+++ service-loans/src/main/java/com/capitaworld/service/loans/service/teaser/primaryview/WorkingCapitalPrimaryViewService.java
@@ -6,5 +6,5 @@ import com.capitaworld.service.loans.model.teaser.primaryview.WorkingCapitalPrim
  * Created by dhaval on 19-May-17.
  */
 public interface WorkingCapitalPrimaryViewService {
-    public WorkingCapitalPrimaryViewResponse getWorkingCapitalPrimaryViewDetails(Long toApplicationId,Long userType,Long fundProviderUserId);
+    public WorkingCapitalPrimaryViewResponse getWorkingCapitalPrimaryViewDetails(Long toApplicationId,Integer userType,Long fundProviderUserId);
 }
